+++ src/main/java/com/github/javafaker/service/FakeValuesService.java
@@ -1,8 +1,6 @@
 package com.github.javafaker.service;
 
 import com.github.javafaker.Resolver;
-import org.apache.commons.lang.StringUtils;
-import org.apache.commons.lang.WordUtils;
 import org.yaml.snakeyaml.Yaml;
 
 import java.io.InputStream;
@@ -12,30 +10,36 @@ import java.util.Map;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-public class FakeValuesService {
+public class FakeValuesService implements FakeValuesServiceInterface {
     private static final char[] METHOD_NAME_DELIMITERS = {'_'};
     private final Map<String, Object> fakeValuesMap;
     private final RandomService randomService;
 
     @SuppressWarnings({"unchecked", "rawtypes"})
     public FakeValuesService(Locale locale, RandomService randomService) {
-        String languageCode = locale.getLanguage();
-        final InputStream stream = findStream("/" + languageCode + ".yml");
+        String filename = locale.getLanguage();
+        InputStream stream = findStream(filename);
+        if (stream == null) {
+            filename = filename + "-" + locale.getCountry();
+            stream = findStream(filename);
+        }
         if (stream == null) {
             throw new LocaleDoesNotExistException(String.format("%s could not be found, does not have a corresponding yaml file", locale));
         }
+
         Map valuesMap = (Map) new Yaml().load(stream);
-        valuesMap = (Map) valuesMap.get(languageCode);
+        valuesMap = (Map) valuesMap.get(filename);
         fakeValuesMap = (Map<String, Object>) valuesMap.get("faker");
         this.randomService = randomService;
     }
 
     private InputStream findStream(String filename) {
-        InputStream streamOnClass = getClass().getResourceAsStream(filename);
+        String filenameWithExtension =  "/" + filename + ".yml";
+        InputStream streamOnClass = getClass().getResourceAsStream(filenameWithExtension);
         if (streamOnClass != null) {
             return streamOnClass;
         }
-        return getClass().getClassLoader().getResourceAsStream(filename);
+        return getClass().getClassLoader().getResourceAsStream(filenameWithExtension);
     }
 
     /**
@@ -55,8 +59,33 @@ public class FakeValuesService {
      * @param key
      * @return
      */
+    // TODO test
     public String fetchString(String key) {
         return (String) fetch(key);
+    }
+
+    /**
+     * Safely fetches a key.
+     *
+     * If the value is null, it will return an empty string.
+     *
+     * If it is a list, it will assume it is a list of strings and select a random value from it.
+     *
+     * Otherwise it will just return the value as a string.
+     *
+     * @param key
+     * @return
+     */
+    @SuppressWarnings({"unchecked", "rawtypes"})
+    public String safeFetch(String key) {
+        Object o = fetchObject(key);
+        if (o == null) return "";
+        if (o instanceof List) {
+            List<String> values = (List<String>) o;
+            return values.get(randomService.nextInt(values.size()));
+        } else {
+            return (String) o;
+        }
     }
 
     /**
@@ -66,6 +95,7 @@ public class FakeValuesService {
      *            dot. E.g. name.first_name
      * @return
      */
+    // TODO test
     @SuppressWarnings({"unchecked", "rawtypes"})
     public Object fetchObject(String key) {
         String[] path = key.split("\\.");
@@ -74,42 +104,6 @@ public class FakeValuesService {
             currentValue = ((Map<String, Object>) currentValue).get(pathSection);
         }
         return currentValue;
-    }
-
-    /**
-     * A property that is composed of other properties.
-     * <p/>
-     * It firstly fetches the formatKey using {@link #fetch(String)}. It will
-     * proceed to convert the returned properties from the {@link #fetch(String)}
-     * method to a methodName and invoke this method against the object passed in.
-     * Finally, concatenation occurs with the return values of the methods
-     * using the joiner parameter as a separator.
-     *
-     * @param formatKey
-     * @param joiner
-     * @param objectToInvokeMethodOn
-     * @return
-     */
-    @SuppressWarnings({"unchecked", "rawtypes"})
-    public String composite(String formatKey, String joiner, Object objectToInvokeMethodOn) {
-        List<String> format = (List<String>) fetch(formatKey);
-
-        String[] parts = new String[format.size()];
-        for (int i = 0; i < parts.length; i++) {
-            // remove leading colon
-            String methodName = format.get(i).substring(1);
-            // convert to camel case
-            methodName = WordUtils.capitalizeFully(methodName, METHOD_NAME_DELIMITERS).replaceAll("_", "");
-            methodName = methodName.substring(0, 1).toLowerCase() + methodName.substring(1);
-
-            try {
-                parts[i] = (String) objectToInvokeMethodOn.getClass().getMethod(methodName, (Class[]) null).invoke(objectToInvokeMethodOn);
-            } catch (Exception e) {
-                throw new RuntimeException(e);
-            }
-        }
-
-        return StringUtils.join(parts, joiner);
     }
 
     /**